+++ android/app/src/main/kotlin/com/openaudioanalyzer/oaa/MainActivity.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package dev.openaudioanalyzer.oaa
+package com.openaudioanalyzer.oaa
 
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine